ModelNodeOutputType
