+++ lang/model_tree/mod.rs
@@ -8,6 +8,7 @@ pub mod iter;
 pub mod metadata;
 pub mod model_node;
 pub mod model_node_builder;
+pub mod model_node_output;
 pub mod model_nodes;
 pub mod object;
 pub mod transformation;
@@ -17,6 +18,7 @@ pub use iter::*;
 pub use metadata::*;
 pub use model_node::*;
 pub use model_node_builder::*;
+pub use model_node_output::*;
 pub use model_nodes::*;
 pub use object::*;
 pub use transformation::*;
